requirements.txt
streamlit
biopython
requests
py3dmol